separating titan_cli from titan dataset file
