+++ py4cast/datasets/titan/__init__.py
@@ -12,10 +12,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 import torch
 import tqdm
-import typer
 import xarray as xr
 from skimage.transform import resize
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import DataLoader
 
 from py4cast.datasets.base import (
     DatasetABC,
@@ -37,11 +36,10 @@ from py4cast.forcingutils import generate_toa_radiation_forcing, get_year_hour_f
 from py4cast.plots import DomainInfo
 from py4cast.utils import merge_dicts
 
-app = typer.Typer()
 
 class TitanAccessor(DataAccessor):
 
-    def get_weight_per_lvl(
+    def get_weight_per_level(
         level: int,
         level_type: Literal["isobaricInhPa", "heightAboveGround", "surface", "meanSea"],
     ):
@@ -50,11 +48,9 @@ class TitanAccessor(DataAccessor):
         else:
             return 2.0
 
-
     #############################################################
     #                            GRID                           #
     #############################################################
-
 
     def load_grid_info(name: str) -> GridConfig:
         if name not in ["PAAROME_1S100", "PAAROME_1S40"]:
@@ -76,12 +72,11 @@ class TitanAccessor(DataAccessor):
         return grid_conf
 
     def get_grid_coords(param: Param) -> List[int]:
-       return METADATA["GRIDS"][param.grid.name]["extent"]
+        return METADATA["GRIDS"][param.grid.name]["extent"]
 
     #############################################################
     #                              PARAMS                       #
     #############################################################
-
 
     def load_param_info(name: str) -> ParamConfig:
         info = METADATA["WEATHER_PARAMS"][name]
@@ -96,6 +91,10 @@ class TitanAccessor(DataAccessor):
                 "Parameter native grid must be in ['PAAROME_1S100', 'PAAROME_1S40', 'PA_01D']"
             )
         return ParamConfig(unit, level_type, long_name, grid, grib_name, grib_param)
+
+    #############################################################
+    #                              LOADING                      #
+    #############################################################
 
     def get_filepath(
         ds_name: str,
@@ -116,22 +115,67 @@ class TitanAccessor(DataAccessor):
             filename = f"{param.name}_{param.level}_{param.level_type}.npy"
             return npy_path / date.strftime(FORMATSTR) / filename
 
+    def get_dataset_path(name: str, grid: Grid):
+        str_subdomain = "-".join([str(i) for i in grid.subdomain])
+        subdataset_name = f"{name}_{grid.name}_{str_subdomain}"
+        return SCRATCH_PATH / "subdatasets" / subdataset_name
 
-def process_sample_dataset(ds_name: str, date: dt.datetime, params: List[Param]):
-    """Saves each 2D parameter data of the given date as one NPY file."""
-    for param in params:
-        dest_file = get_filepath(ds_name, param, date, "npy")
-        dest_file.parent.mkdir(exist_ok=True)
-        if not dest_file.exists():
-            try:
-                arr = load_data_from_disk(ds_name, param, date, "grib")
-                np.save(dest_file, arr)
-            except Exception as e:
-                print(e)
-                print(
-                    f"WARNING: Could not load grib data {param.name} {param.level} {date}. Skipping sample."
-                )
-                break
+    def load_data_from_disk(
+        ds_name: str,
+        param: Param,
+        date: dt.datetime,
+        file_format: Literal["npy", "grib"] = "grib",
+    ) -> np.ndarray:
+        """
+        Function to load invidiual parameter and lead time from a file stored in disk
+        """
+        data_path = get_filepath(ds_name, param, date, file_format)
+        if file_format == "grib":
+            arr, lons, lats = load_data_grib(param, data_path)
+            arr = fit_to_grid(arr, lons, lats)
+        else:
+            arr = np.load(data_path)
+
+        subdomain = param.grid.subdomain
+        arr = arr[subdomain[0] : subdomain[1], subdomain[2] : subdomain[3]]
+        if file_format == "grib":
+            arr = arr[::-1]
+        return arr  # invert latitude
+
+    def get_param_tensor(
+        param: Param,
+        stats: Stats,
+        dates: List[dt.datetime],
+        settings: Settings,
+        no_standardize: bool = False,
+    ) -> torch.tensor:
+        """
+        Fetch data on disk fo the given parameter and all involved dates
+        Unless specified, normalize the samples with parameter-specific constants
+        returns a tensor
+        """
+        arrays = [
+            load_data_from_disk(
+                settings.dataset_name, param, date, settings.file_format
+            )
+            for date in dates
+        ]
+        arr = np.stack(arrays)
+        # Extend dimension to match 3D (level dimension)
+        if len(arr.shape) != 4:
+            arr = np.expand_dims(arr, axis=1)
+        arr = np.transpose(arr, axes=[0, 2, 3, 1])  # shape = (steps, lvl, x, y)
+        if settings.standardize and not no_standardize:
+            name = param.parameter_short_name
+            means = np.asarray(stats[name]["mean"])
+            std = np.asarray(stats[name]["std"])
+            arr = (arr - means) / std
+        return torch.from_numpy(arr)
+
+
+############################################################
+#                   HELPER FUNCTIONS for TITAN             #
+############################################################
 
 
 def fit_to_grid(
@@ -179,106 +223,12 @@ def load_data_grib(param: Param, path: Path) -> np.ndarray:
     return arr, lons, lats
 
 
-def load_data_from_disk(
-    ds_name: str,
-    param: Param,
-    date: dt.datetime,
-    file_format: Literal["npy", "grib"] = "grib",
-) -> np.ndarray :
-    """
-    Function to load invidiual parameter and lead time from a file stored in disk
-    """
-    data_path = get_filepath(ds_name, param, date, file_format)
-    if file_format == "grib":
-        arr, lons, lats = load_data_grib(param, data_path)
-        arr = fit_to_grid(arr, lons, lats)
-    else:
-        arr = np.load(data_path)
-
-    subdomain = param.grid.subdomain
-    arr = arr[subdomain[0] : subdomain[1], subdomain[2] : subdomain[3]]
-    if file_format == "grib":
-        arr = arr[::-1]
-    return arr  # invert latitude
-
-
-def exists(
-    ds_name: str,
-    param: Param,
-    date: dt.datetime,
-    file_format: Literal["npy", "grib"] = "grib",
-) -> bool:
-    filepath = get_filepath(ds_name, param, date, file_format)
-    return filepath.exists()
-
-
-def get_param_tensor(
-    param: Param,
-    stats: Stats,
-    dates: List[dt.datetime],
-    settings: Settings,
-    no_standardize: bool = False,
-) -> torch.tensor:
-    """
-    Fetch data on disk fo the given parameter and all involved dates
-    Unless specified, normalize the samples with parameter-specific constants
-    returns a tensor
-    """
-    arrays = [
-        load_data_from_disk(settings.dataset_name, param, date, settings.file_format)
-        for date in dates
-    ]
-    arr = np.stack(arrays)
-    # Extend dimension to match 3D (level dimension)
-    if len(arr.shape) != 4:
-        arr = np.expand_dims(arr, axis=1)
-    arr = np.transpose(arr, axes=[0, 2, 3, 1])  # shape = (steps, lvl, x, y)
-    if settings.standardize and not no_standardize:
-        name = param.parameter_short_name
-        means = np.asarray(stats[name]["mean"])
-        std = np.asarray(stats[name]["std"])
-        arr = (arr - means) / std
-    return torch.from_numpy(arr)
-
-
-def generate_forcings(
-    date: dt.datetime, output_terms: Tuple[float], grid: Grid
-) -> List[NamedTensor]:
-    """
-    Generate all the forcing in this function.
-    Return a list of NamedTensor.
-    """
-    lforcings = []
-    time_forcing = NamedTensor(  # doy : day_of_year
-        feature_names=["cos_hour", "sin_hour", "cos_doy", "sin_doy"],
-        tensor=get_year_hour_forcing(date, output_terms).type(torch.float32),
-        names=["timestep", "features"],
-    )
-    solar_forcing = NamedTensor(
-        feature_names=["toa_radiation"],
-        tensor=generate_toa_radiation_forcing(
-            grid.lat, grid.lon, date, output_terms
-        ).type(torch.float32),
-        names=["timestep", "lat", "lon", "features"],
-    )
-    lforcings.append(time_forcing)
-    lforcings.append(solar_forcing)
-
-    return lforcings
-
-
 #############################################################
 #                            DATASET                        #
 #############################################################
 
 
-def get_dataset_path(name: str, grid: Grid):
-    str_subdomain = "-".join([str(i) for i in grid.subdomain])
-    subdataset_name = f"{name}_{grid.name}_{str_subdomain}"
-    return SCRATCH_PATH / "subdatasets" / subdataset_name
-
-
-class TitanDataset(DatasetABC, Dataset):
+class TitanDataset(DatasetABC):
     # Si on doit travailler avec plusieurs grilles, on fera un super dataset qui contient
     # plusieurs datasets chacun sur une seule grille
     def __init__(
@@ -289,22 +239,7 @@ class TitanDataset(DatasetABC, Dataset):
         params: List[Param],
         settings: Settings,
     ):
-        self.name = name
-        self.grid = grid
-        if grid.name not in ["PAAROME_1S100", "PAAROME_1S40"]:
-            raise NotImplementedError(
-                "Grid must be in ['PAAROME_1S100', 'PAAROME_1S40']"
-            )
-        self.period = period
-        self.params = params
-        self.settings = settings
-        self.shuffle = self.period.name == "train"
-        self.cache_dir.mkdir(parents=True, exist_ok=True)
-
-        n_input, n_pred = self.settings.num_input_steps, self.settings.num_pred_steps
-        filename = f"valid_samples_{self.period.name}_{n_input}_{n_pred}.txt"
-        self.valid_samples_file = self.cache_dir / filename
-
+        super().__init__(self, name, grid, period, params, settings, TitanAccessor)
 
     @cached_property
     def sample_list(self):
@@ -333,124 +268,3 @@ class TitanDataset(DatasetABC, Dataset):
                     samples.append(sample)
         print(f"--> All {len(samples)} {self.period.name} samples are now defined")
         return samples
-
-
-@app.command()
-def prepare(
-    path_config: Path = DEFAULT_CONFIG,
-    num_input_steps: int = 1,
-    num_pred_steps_train: int = 1,
-    num_pred_steps_val_test: int = 1,
-    convert_grib2npy: bool = False,
-    compute_stats: bool = True,
-    write_valid_samples_list: bool = True,
-):
-    """Prepares Titan dataset for training.
-    This command will:
-        - create all needed folders
-        - convert gribs to npy and rescale data to the wanted grid
-        - establish a list of valid samples for each set
-        - computes statistics on all weather parameters."""
-    print("--> Preparing Titan Dataset...")
-
-    print("Load dataset configuration...")
-    with open(path_config, "r") as fp:
-        conf = json.load(fp)
-
-    print("Creating folders...")
-    train_ds, valid_ds, test_ds = TitanDataset.from_dict(
-        path_config.stem,
-        conf,
-        num_input_steps,
-        num_pred_steps_train,
-        num_pred_steps_val_test,
-    )
-    train_ds.cache_dir.mkdir(exist_ok=True)
-    data_dir = train_ds.cache_dir / "data"
-    data_dir.mkdir(exist_ok=True)
-    print(f"Dataset will be saved in {train_ds.cache_dir}")
-
-    if convert_grib2npy:
-        print("Converting gribs to npy...")
-        param_list = get_param_list(
-            conf, train_ds.grid, load_param_info, get_weight_per_lvl
-        )
-        sum_dates = (
-            list(train_ds.period.date_list)
-            + list(valid_ds.period.date_list)
-            + list(test_ds.period.date_list)
-        )
-        dates = sorted(list(set(sum_dates)))
-        for date in tqdm.tqdm(dates):
-            process_sample_dataset(date, param_list)
-        print("Done!")
-
-    conf["settings"]["standardize"] = False
-    train_ds, valid_ds, test_ds = TitanDataset.from_dict(
-        path_config.stem,
-        conf,
-        num_input_steps,
-        num_pred_steps_train,
-        num_pred_steps_val_test,
-    )
-    if compute_stats:
-        print("Computing stats on each parameter...")
-        train_ds.compute_parameters_stats()
-    if write_valid_samples_list:
-        train_ds.write_list_valid_samples()
-        valid_ds.write_list_valid_samples()
-        test_ds.write_list_valid_samples()
-
-    if compute_stats:
-        print("Computing time stats on each parameters, between 2 timesteps...")
-        conf["settings"]["standardize"] = True
-        train_ds, valid_ds, test_ds = TitanDataset.from_dict(
-            path_config.stem,
-            conf,
-            num_input_steps,
-            num_pred_steps_train,
-            num_pred_steps_val_test,
-        )
-        train_ds.compute_time_step_stats()
-
-
-@app.command()
-def describe(path_config: Path = DEFAULT_CONFIG):
-    """Describes Titan."""
-    train_ds, _, _ = TitanDataset.from_json(path_config, 2, 1, 5)
-    train_ds.dataset_info.summary()
-    print("Len dataset : ", len(train_ds))
-    print("First Item description :")
-    print(train_ds[0])
-
-
-@app.command()
-def plot(path_config: Path = DEFAULT_CONFIG):
-    """Plots a png and a gif for one sample."""
-    train_ds, _, _ = TitanDataset.from_json(path_config, 2, 1, 5)
-    print("Plot gif of one sample...")
-    sample = train_ds.sample_list[0]
-    sample.plot_gif("test.gif")
-    print("Plot png for one step of sample...")
-    item = sample.load(no_standardize=True)
-    sample.plot(item, 0, "test.png")
-
-
-@app.command()
-def speedtest(path_config: Path = DEFAULT_CONFIG, n_iter: int = 5):
-    """Makes a loading speed test."""
-    train_ds, _, _ = TitanDataset.from_json(path_config, 2, 1, 5)
-    data_iter = iter(train_ds.torch_dataloader())
-    print("Dataset file_format: ", train_ds.settings.file_format)
-    print("Speed test:")
-    start_time = time.time()
-    for _ in tqdm.trange(n_iter, desc="Loading samples"):
-        next(data_iter)
-    delta = time.time() - start_time
-    print("Elapsed time : ", delta)
-    speed = n_iter / delta
-    print(f"Loading speed: {round(speed, 3)} batch(s)/sec")
-
-
-if __name__ == "__main__":
-    app()
